created new folders to organize components and updated imports
